First attempt at adding the Solr+Ontopia prototype to the sandbox. The current check-in contains a Solr index directories with two empty cores (read: indexes).
The Maven project contains a single Java class that can be run to produce a
Solr index from a topic map. To to this you need to add a couple of topic
types with instances to your topic map. I'll try to provide a sample topic
map soon, but as the topic map that was used for this prototype belongs to a
customer I cannot add that one here. If you're interested enough you should
be able to figure this out by looking at the presentation and the
SolrIndexer.java class.

That's all for now.