Lots of improvements to html and css code. Added support for deploying ontopoly web app to running tomcat instance using maven cargo plugin.
To make it possible to deploy Ontopia needs to be build with the standard
Tomcat web applications. Note that they are removed by default because of
some security conserns. A distribution that includes those apps can be built
with the following command in the source root directory:

  ant -Dkeep.tomcat.webapps=true dist.ontopia

To build Ontopoly and deploy it issue this command inside the ontopoly maven
module:

  mvn -Dmaven.test.skip clean install cargo:redeploy